🔨 extract shared bespoke components and hooks
