+++ bespoke/components/ChartFooter/ChartFooter.scss
@@ -1,0 +1,43 @@
+.chart-footer {
+    font-size: 0.8125rem;
+    margin-top: 8px;
+
+    strong {
+        font-weight: 600;
+    }
+
+    .chart-footer__note {
+        font-size: 0.75rem;
+        margin-top: 4px;
+    }
+
+    .chart-footer__source-and-cc {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        flex-wrap: wrap;
+
+        column-gap: 16px;
+        row-gap: 2px;
+    }
+
+    .chart-footer__cc {
+        background: none;
+        border: none;
+        cursor: pointer;
+        text-decoration: none;
+        color: inherit;
+    }
+
+    @at-root .chart-footer__tooltip {
+        background: white;
+        color: $gray-100;
+        border-radius: 4px;
+        box-shadow: 0px 4px 40px rgba(0, 0, 0, 0.2);
+        padding: 8px 12px;
+        font-size: 0.8125rem;
+        max-width: 300px;
+        line-height: 1.5;
+        font-family: $sans-serif-font-stack;
+    }
+}
