+++ bespoke/components/TimeSlider/TimeSlider.scss
@@ -1,0 +1,127 @@
+.time-slider {
+    $timelineHeight: 32px;
+
+    $handle-diameter: 16px;
+    $default-knob: $gray-60;
+    $active-knob: $blue-50;
+    $light-timeline: $gray-20;
+    $active-timeline: $blue-30;
+    $hover-timeline: $default-knob;
+    $light-fill: $gray-10;
+    $hover-fill: $gray-20;
+    $active-fill: $blue-20;
+
+    width: 100%;
+    height: $timelineHeight;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 1em;
+
+    > div {
+        margin-right: 0.5em;
+        &:last-child {
+            margin-right: 0;
+        }
+    }
+
+    .date {
+        height: 100%;
+        font-size: 13px;
+        font-weight: 400;
+        color: $gray-90;
+        padding: 0 8px;
+        border-radius: 4px;
+        display: grid;
+        place-content: center;
+        margin-right: 0;
+        background-color: $light-fill;
+        user-select: none;
+        border: none;
+        cursor: pointer;
+
+        &:hover {
+            background: $hover-fill;
+        }
+
+        &:active {
+            background: $active-fill;
+            color: $blue-90;
+        }
+    }
+
+    .slider {
+        height: 2px;
+        flex-grow: 1;
+        background: $light-timeline;
+        position: relative;
+        display: flex;
+        align-items: center;
+        margin-left: $handle-diameter;
+        margin-right: $handle-diameter;
+        padding: 12px 0;
+        box-sizing: content-box !important;
+        background-clip: content-box !important;
+    }
+
+    .handle {
+        $handleSidePadding: 6px;
+
+        position: absolute;
+        z-index: 1;
+        margin-left: (-$handle-diameter * 0.5) - $handleSidePadding;
+        cursor: col-resize;
+        padding: 10px $handleSidePadding;
+        box-sizing: content-box !important;
+        background-clip: content-box !important;
+
+        > .icon {
+            height: $handle-diameter;
+            width: $handle-diameter;
+            border-radius: 100%;
+            background: $default-knob;
+            border: 2px solid #fff;
+            z-index: 1;
+            pointer-events: none;
+            transform: scale(1);
+            transition: transform 0.1s ease-out;
+        }
+    }
+
+    &.hover {
+        .handle > .icon {
+            transform: scale(1.3);
+            background: $active-knob;
+        }
+    }
+}
+
+.time-slider__tooltip {
+    position: absolute;
+    top: -28px;
+    left: 50%;
+    transform: translateX(-50%);
+    background-color: $gray-90;
+    color: white;
+    padding: 6px 8px;
+    border-radius: 4px;
+    font-size: 12px;
+    line-height: 16px;
+    white-space: nowrap;
+    z-index: 100;
+    pointer-events: none;
+
+    // Tooltip arrow
+    &::after {
+        content: "";
+        position: absolute;
+        top: 100%;
+        left: 50%;
+        transform: translateX(-50%);
+        width: 0;
+        height: 0;
+        border-left: 4px solid transparent;
+        border-right: 4px solid transparent;
+        border-top: 4px solid $gray-90;
+    }
+}
